Add files via upload
Updating with comments the scraping_from_lista_firme
Creating the captcha that needs some improvements in order to collect data from balance sheet. It has to be decide whether we go on listafirme or mfinante as the number of features vary and int may influence the performance of the model.
the companies_url.RData is the workspace which contains all the links to companies and the links to pages with companies.