Almos fully-independent version
The script can scrap data. After every 500 companies, it sleeps for 5 minuters, in this way, I am trying to avoid non-human behaviour detection, doesn't seem that it works very well, but I will check it later. Once the website detects the behaviour it ask to pass some dynamic captchas, currently unable to solve them, have to go deeper in the field of captcha or to check with chenging the IP address. Currently, there is data for 20 000 companies and it takes to many resources.