Fixing the duplicates in the list
The list of companies was populated with duplicates. Starting with the page 2000, all companies on listafirme are the same as before, I had to find why do I have duplicates and also to remove those, currently working fine, but I reduced the number of companies to 100 000 instead of 842 000 :( :( :( 
I hope it will be enough.